requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
solana==0.32.0
solders>=0.18.0
websockets==12.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pandas==2.1.4
openpyxl==3.1.2
python-dotenv>=1.0.0